Minor refactor to allow access to certain quantities
Previously, the GeometryEngine did not ever return the normalizing
constant of the torsion distribution (since it was immediately used to
normalize the discrete pdf). Since it is now required, this commit
moves most of the innards of the function that used to return logpmf to
a new function that does not normalize, then leaves the normalization up
to a function that bears the same name as the old one.